+++ countlyCommon/TestingRelated/RequestTestCases.cs
@@ -64,7 +64,7 @@ namespace TestProject_common
 
             Assert.Equal("YOUR_APP_KEY", baseParams["app_key"]);
             Assert.Equal("test device id", baseParams["device_id"]);
-            Assert.Equal("24.1.0", baseParams["sdk_version"]);
+            Assert.Equal("24.1.1", baseParams["sdk_version"]);
             Assert.Equal(0, baseParams["t"]);
             Assert.Equal("1.0", baseParams["av"]);